Refactor StatsPanel to use factory classes
