+++ src/main/java/seedu/moolah/model/statistics/Statistics.java
@@ -1,5 +1,7 @@
 package seedu.moolah.model.statistics;
 
+import seedu.moolah.ui.statistics.StatisticsRegionFactory;
+
 /**
  * The API of a Statistics object
  */
@@ -10,5 +12,7 @@ public interface Statistics {
     void setTitle(String title);
 
     void populateData();
+
+    StatisticsRegionFactory createFactory();
 
 }
